+++ activerecord/test/cases/validations/numericality_validation_test.rb
@@ -12,10 +12,10 @@ class NumericalityValidationTest < ActiveRecord::TestCase
 
   def test_column_with_precision
     model_class.validates_numericality_of(
-      :bank_balance, equal_to: 10_000_000.12
+      :unscaled_bank_balance, equal_to: 10_000_000.12
     )
 
-    subject = model_class.new(bank_balance: 10_000_000.121)
+    subject = model_class.new(unscaled_bank_balance: 10_000_000.121)
 
     assert_predicate subject, :valid?
   end
@@ -28,6 +28,16 @@ class NumericalityValidationTest < ActiveRecord::TestCase
     subject = model_class.new(decimal_number_big_precision: 10_000_000.3)
 
     assert_predicate subject, :valid?
+  end
+
+  def test_column_with_scale
+    model_class.validates_numericality_of(
+      :bank_balance, greater_than: 10
+    )
+
+    subject = model_class.new(bank_balance: 10.001)
+
+    assert_not_predicate subject, :valid?
   end
 
   def test_no_column_precision
@@ -69,5 +79,27 @@ class NumericalityValidationTest < ActiveRecord::TestCase
     subject = klass.new(bank_balance: 10_000_000.12)
 
     assert_predicate(subject, :valid?)
+  end
+
+  def test_virtual_attribute_with_precision
+    model_class.attribute(:virtual_decimal_number, :decimal, precision: 5)
+    model_class.validates_numericality_of(
+      :virtual_decimal_number, equal_to: 123.45
+    )
+
+    subject = model_class.new(virtual_decimal_number: 123.455)
+
+    assert_predicate subject, :valid?
+  end
+
+  def test_virtual_attribute_with_scale
+    model_class.attribute(:virtual_decimal_number, :decimal, scale: 2)
+    model_class.validates_numericality_of(
+      :virtual_decimal_number, greater_than: 1
+    )
+
+    subject = model_class.new(virtual_decimal_number: 1.001)
+
+    assert_not_predicate subject, :valid?
   end
 end